Move FileMoveIcon to view package.
Move it to view package because even though it is a kind of model, it's a model that is used purely for displaying things.  It uses SWT classes, and it is used only by classes in the view package (which no longer need an explicit import for it, now that it's in their package.)  Clean up whitespace/newlines around the actual enum values, and use value from Constants in place of hard-coded "res".
